pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "phelipi", email = "ggizmeuamigo@gmail.com"},
]
dependencies = [
    "django>=4.2.4",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.0.0",
    "setuptools>=68.0.0",
]
requires-python = ">=3.10"
license = {text = "MIT"}
